fixup: move disruption form to parent disruption live view
