+++ lib/arrow_web/live/disruption_v2_live/disruption_v2_view_live.html.heex
@@ -8,11 +8,9 @@
 
 <h2>Disruption</h2>
 
-<.live_component
-  module={ArrowWeb.DisruptionFormComponent}
+<.disruption_form
   id="disruption_v2_form"
   form={@form}
   disruption_v2={@disruption_v2}
   icon_paths={@icon_paths}
-  action={@form_action}
 />
